refactor: move global-options to data-dir/settings subpackage
Moves packages/core-infra/src/global-options/ to
packages/core-infra/src/data-dir/settings/ and renames the subpath
export from @dotobokuri/core-infra/global-options to
@dotobokuri/core-infra/data-dir/settings. Updates exports/typesVersions
in package.json and tsup entry accordingly. Symbol names (GlobalOptionsService,
GlobalOptionsData, etc.) are preserved. Test files renamed to
data-dir-settings-*.test.ts with updated describe strings and imports.
