+++ packages/core-infra/src/index.ts
@@ -1,5 +1,5 @@
 import { createAuthService, DEFAULT_AUTH_PATH } from "./auth/auth-storage.js";
-import { createGlobalOptionsService, type GlobalOptionsService } from "./global-options/index.js";
+import { createGlobalOptionsService, type GlobalOptionsService } from "./data-dir/settings/index.js";
 import type { AuthService } from "./auth/types.js";
 
 export interface InfraServices {
@@ -14,7 +14,7 @@ export interface InfraServicesDeps {
 export * from "./auth/index.js";
 export * from "./data-dir/index.js";
 export * from "./fs-store/index.js";
-export * from "./global-options/index.js";
+export * from "./data-dir/settings/index.js";
 
 export function createInfraServices(deps: InfraServicesDeps = {}): InfraServices {
   const globalOptionsService = createGlobalOptionsService();
